refactor: move PostSingleLayour and PostListItem into SharedComponents
