+++ src/components/SharedComponents/PostSingleLayout/PostSingleLayout.tsx
@@ -1,10 +1,9 @@
 import { FC, ReactNode, useId } from "react";
 
-import PostCategoryList from "./PostCategoryList";
-import { PostCategoryPage } from "./PostCategoryList/PostCategoryList";
-import usePostCategoryList from "./PostCategoryList/usePostCategoryList";
-import BlogHeader from "./PostHeader/PostHeader";
-
+import PostCategoryList from "@/components/Posts/PostCategoryList";
+import { PostCategoryPage } from "@/components/Posts/PostCategoryList/PostCategoryList";
+import usePostCategoryList from "@/components/Posts/PostCategoryList/usePostCategoryList";
+import BlogHeader from "@/components/Posts/PostHeader/PostHeader";
 import { WebinarSinglePageProps } from "@/pages/webinars/[webinarSlug]";
 import { BlogSinglePageProps } from "@/pages/blog/[blogSlug]";
 import theme from "@/styles/theme";
